Add files via upload
reverse shell.py is a code that I created to simulate a backdoor to the connected server.  After we input the port number, set the listen to true, and set command to true, we now have remote access to the victim and can run any command from the victim cmd and send the command result to our pc.